requirements.txt
flask
flask-smorest
flask-sqlalchemy
flask-jwt-extended
passlib
jsonify